Added new OrientBatchGraph implementation that avoid transactions and use the massive insert intent
