+++ blueprints-orient-graph/src/main/java/com/tinkerpop/blueprints/impls/orient/OrientElement.java
@@ -1,5 +1,8 @@
 package com.tinkerpop.blueprints.impls.orient;
 
+import java.util.HashSet;
+import java.util.Set;
+
 import com.orientechnologies.orient.core.db.record.OIdentifiable;
 import com.orientechnologies.orient.core.db.record.ORecordElement.STATUS;
 import com.orientechnologies.orient.core.exception.OSerializationException;
@@ -14,18 +17,15 @@ import com.tinkerpop.blueprints.util.ElementHelper;
 import com.tinkerpop.blueprints.util.ExceptionFactory;
 import com.tinkerpop.blueprints.util.StringFactory;
 
-import java.util.HashSet;
-import java.util.Set;
-
 /**
  * @author Luca Garulli (http://www.orientechnologies.com)
  */
 abstract class OrientElement implements Element, OSerializableStream, OIdentifiable {
 
-    protected final OrientGraph graph;
+    protected final OrientBaseGraph graph;
     protected final ODocument rawElement;
 
-    protected OrientElement(final OrientGraph rawGraph, final ODocument rawElement) {
+    protected OrientElement(final OrientBaseGraph rawGraph, final ODocument rawElement) {
         this.graph = rawGraph;
         this.rawElement = rawElement;
     }
