+++ blueprints-orient-graph/src/main/java/com/tinkerpop/blueprints/impls/orient/OrientElementScanIterable.java
@@ -1,5 +1,7 @@
 package com.tinkerpop.blueprints.impls.orient;
 
+import java.util.Iterator;
+
 import com.orientechnologies.orient.core.db.graph.OGraphDatabase;
 import com.orientechnologies.orient.core.db.record.ODatabaseRecordAbstract;
 import com.orientechnologies.orient.core.iterator.ORecordIteratorClass;
@@ -8,17 +10,15 @@ import com.tinkerpop.blueprints.CloseableIterable;
 import com.tinkerpop.blueprints.Element;
 import com.tinkerpop.blueprints.Vertex;
 
-import java.util.Iterator;
-
 /**
  * @author Luca Garulli (http://www.orientechnologies.com)
  */
 class OrientElementScanIterable<T extends Element> implements CloseableIterable<T> {
     private final Class<T> elementClass;
-    private final OrientGraph graph;
+    private final OrientBaseGraph graph;
     private final boolean polymorphic;
 
-    public OrientElementScanIterable(final OrientGraph graph, Class<T> elementClass, final boolean polymorphic) {
+    public OrientElementScanIterable(final OrientBaseGraph graph, Class<T> elementClass, final boolean polymorphic) {
         this.graph = graph;
         this.elementClass = elementClass;
         this.polymorphic = polymorphic;
