+++ blueprints-orient-graph/src/main/java/com/tinkerpop/blueprints/impls/orient/OrientVertex.java
@@ -1,5 +1,12 @@
 package com.tinkerpop.blueprints.impls.orient;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
 import com.orientechnologies.orient.core.db.graph.OGraphDatabase;
 import com.orientechnologies.orient.core.db.record.OIdentifiable;
 import com.orientechnologies.orient.core.record.impl.ODocument;
@@ -12,13 +19,6 @@ import com.tinkerpop.blueprints.util.MultiIterable;
 import com.tinkerpop.blueprints.util.StringFactory;
 import com.tinkerpop.blueprints.util.VerticesFromEdgesIterable;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
 
 /**
  * @author Luca Garulli (http://www.orientechnologies.com)
@@ -29,7 +29,7 @@ public class OrientVertex extends OrientElement implements Vertex {
         super(null, new ODocument());
     }
 
-    public OrientVertex(final OrientGraph rawGraph, final ODocument rawVertex) {
+    public OrientVertex(final OrientBaseGraph rawGraph, final ODocument rawVertex) {
         super(rawGraph, rawVertex);
     }
